Remove handlers module and implement chat_impl in router.rs
# Conflicts:
#	sgl-router/src/routers/grpc/router.rs
