--- sgl-router/src/routers/grpc/handlers/chat.rs
@@ -1,447 +1,0 @@
-// Chat completion implementation for gRPC Router
-
-use super::super::router::GrpcRouter;
-use crate::grpc::proto;
-use crate::grpc::SglangSchedulerClient;
-use crate::protocols::spec::{ChatCompletionRequest, ChatMessage, StringOrArray, UserMessageContent, ContentPart, ResponseFormat};
-use crate::tokenizer::chat_template::ChatMessage as TokenizerChatMessage;
-use crate::core::Worker;
-use axum::{
-    http::{HeaderMap, StatusCode},
-    response::{IntoResponse, Response},
-};
-use std::sync::Arc;
-use tracing::{debug, error, warn};
-use uuid::Uuid;
-
-// Data structures for processing
-#[derive(Debug)]
-pub struct ProcessedMessages {
-    pub text: String,
-    pub multimodal_inputs: Option<proto::MultimodalInputs>,
-    pub stop_sequences: Option<StringOrArray>,
-}
-
-impl GrpcRouter {
-    /// Main route_chat implementation
-    pub async fn route_chat_impl(
-        &self,
-        _headers: Option<&HeaderMap>,
-        body: &ChatCompletionRequest,
-        model_id: Option<&str>,
-    ) -> Response {
-        debug!("Processing chat completion request for model: {:?}", model_id);
-
-        // Step 1: Validate request
-        if let Err(e) = self.validate_chat_request(body) {
-            warn!("Chat request validation failed: {}", e);
-            return (StatusCode::BAD_REQUEST, e.to_string()).into_response();
-        }
-
-        // Step 2: Process messages and apply chat template
-        let processed_messages = match self.process_chat_messages(body) {
-            Ok(msgs) => msgs,
-            Err(e) => {
-                error!("Failed to process chat messages: {}", e);
-                return (StatusCode::BAD_REQUEST, e.to_string()).into_response();
-            }
-        };
-
-        // Step 3: Tokenize the processed text
-        let encoding = match self.tokenizer.encode(&processed_messages.text) {
-            Ok(encoding) => encoding,
-            Err(e) => {
-                error!("Tokenization failed: {}", e);
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    format!("Tokenization failed: {}", e),
-                )
-                    .into_response();
-            }
-        };
-
-        let token_ids = encoding.token_ids().to_vec();
-
-        debug!("Tokenized {} tokens from input", token_ids.len());
-
-        // Step 4: Build tool constraints if needed
-        let structural_tag = if let Some(tools) = &body.tools {
-            self.generate_tool_constraints(tools, &body.tool_choice, &body.model)
-        } else {
-            None
-        };
-
-        // Step 5: Build SamplingParams for gRPC
-        let sampling_params = self.build_grpc_sampling_params(body, structural_tag);
-
-        // Step 6: Create GenerateRequest
-        let grpc_request = proto::GenerateRequest {
-            request_id: format!("chatcmpl-{}", Uuid::new_v4()),
-            tokenized: Some(proto::TokenizedInput {
-                original_text: processed_messages.text.clone(),
-                input_ids: token_ids.into_iter().map(|id| id as i32).collect(),
-            }),
-            mm_inputs: processed_messages.multimodal_inputs,
-            sampling_params: Some(sampling_params),
-            return_logprob: body.logprobs,
-            logprob_start_len: -1,
-            top_logprobs_num: body.top_logprobs.unwrap_or(0) as i32,
-            return_hidden_states: body.return_hidden_states,
-            ..Default::default()
-        };
-
-        // Step 7: Select worker
-        let worker = match self.select_worker_for_request(model_id, &processed_messages.text) {
-            Some(w) => w,
-            None => {
-                warn!("No available workers for model: {:?}", model_id);
-                return (StatusCode::SERVICE_UNAVAILABLE, "No available workers").into_response();
-            }
-        };
-
-        debug!("Selected worker: {}", worker.url());
-
-        // Step 8: Get gRPC client for worker
-        let client = match self.get_or_create_grpc_client(worker.url()).await {
-            Ok(c) => c,
-            Err(e) => {
-                error!("Failed to get gRPC client: {}", e);
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    format!("Failed to get gRPC client: {}", e),
-                )
-                    .into_response();
-            }
-        };
-
-        // Step 9: Handle streaming vs non-streaming
-        if body.stream {
-            self.handle_streaming_chat(client, grpc_request, body).await
-        } else {
-            self.handle_non_streaming_chat(client, grpc_request, body).await
-        }
-    }
-
-    // ============ Helper Methods ============
-
-    /// Validate chat completion request
-    fn validate_chat_request(&self, request: &ChatCompletionRequest) -> Result<(), String> {
-        // Check if messages are provided
-        if request.messages.is_empty() {
-            return Err("Messages cannot be empty".to_string());
-        }
-
-        // Validate temperature
-        if let Some(temp) = request.temperature {
-            if temp < 0.0 || temp > 2.0 {
-                return Err(format!("Temperature {} is out of range [0, 2]", temp));
-            }
-        }
-
-        // Validate top_p
-        if let Some(top_p) = request.top_p {
-            if top_p < 0.0 || top_p > 1.0 {
-                return Err(format!("top_p {} is out of range [0, 1]", top_p));
-            }
-        }
-
-        // Validate max_tokens
-        if let Some(max_tokens) = request.max_tokens {
-            if max_tokens <= 0 {
-                return Err("max_tokens must be positive".to_string());
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Process chat messages and apply template
-    fn process_chat_messages(
-        &self,
-        request: &ChatCompletionRequest,
-    ) -> Result<ProcessedMessages, String> {
-        // Convert spec ChatMessage enum to tokenizer ChatMessage struct
-        let tokenizer_messages = self.convert_messages_for_tokenizer(&request.messages)?;
-
-        // Apply chat template - need to check the actual interface
-        // For now, create a simple formatted text
-        let formatted_text = self.format_messages_simple(&tokenizer_messages);
-
-        // Extract multimodal inputs if present
-        let multimodal_inputs = self.extract_multimodal_inputs(&request.messages)?;
-
-        Ok(ProcessedMessages {
-            text: formatted_text,
-            multimodal_inputs,
-            stop_sequences: request.stop.clone(),
-        })
-    }
-
-    /// Convert spec ChatMessage enum to tokenizer ChatMessage struct
-    fn convert_messages_for_tokenizer(
-        &self,
-        messages: &[ChatMessage],
-    ) -> Result<Vec<TokenizerChatMessage>, String> {
-        let mut converted = Vec::new();
-
-        for message in messages {
-            let tokenizer_msg = match message {
-                ChatMessage::System { content, .. } => {
-                    TokenizerChatMessage::new("system", content)
-                }
-                ChatMessage::User { content, .. } => {
-                    let text_content = match content {
-                        UserMessageContent::Text(text) => text.clone(),
-                        UserMessageContent::Parts(parts) => {
-                            // Extract text from parts, handling multimodal content
-                            let mut combined_text = String::new();
-                            for part in parts {
-                                match part {
-                                    ContentPart::Text { text } => {
-                                        combined_text.push_str(text);
-                                        combined_text.push('\n');
-                                    }
-                                    ContentPart::ImageUrl { image_url } => {
-                                        // For now, include image URL in text
-                                        combined_text.push_str(&format!("image://{}\n", image_url.url));
-                                    }
-                                }
-                            }
-                            combined_text
-                        }
-                    };
-                    TokenizerChatMessage::new("user", text_content)
-                }
-                ChatMessage::Assistant { content, tool_calls, reasoning_content, .. } => {
-                    let mut assistant_content = String::new();
-
-                    if let Some(reasoning) = reasoning_content {
-                        assistant_content.push_str(&format!("<think>{}</think>\n", reasoning));
-                    }
-
-                    if let Some(content) = content {
-                        assistant_content.push_str(content);
-                    }
-
-                    if let Some(calls) = tool_calls {
-                        // Format tool calls - this would depend on the model's format
-                        for call in calls {
-                            assistant_content.push_str(&format!("\n<tool_call>{}</tool_call>",
-                                serde_json::to_string(call).unwrap_or_default()));
-                        }
-                    }
-
-                    TokenizerChatMessage::new("assistant", assistant_content)
-                }
-                ChatMessage::Tool { content, tool_call_id, .. } => {
-                    TokenizerChatMessage::new("tool", format!("tool_call_id: {}\n{}", tool_call_id, content))
-                }
-                ChatMessage::Function { content, name, .. } => {
-                    TokenizerChatMessage::new("function", format!("function: {}\n{}", name, content))
-                }
-            };
-            converted.push(tokenizer_msg);
-        }
-
-        Ok(converted)
-    }
-
-    /// Simple message formatting fallback
-    fn format_messages_simple(&self, messages: &[TokenizerChatMessage]) -> String {
-        let mut result = String::new();
-        for msg in messages {
-            result.push_str(&format!("<|im_start|>{}\n{}<|im_end|>\n", msg.role, msg.content));
-        }
-        result.push_str("<|im_start|>assistant\n");
-        result
-    }
-
-    /// Extract multimodal inputs from messages
-    fn extract_multimodal_inputs(
-        &self,
-        messages: &[ChatMessage],
-    ) -> Result<Option<proto::MultimodalInputs>, String> {
-        let mut image_urls = Vec::new();
-        let video_urls = Vec::new();
-        let audio_urls = Vec::new();
-
-        for message in messages {
-            if let ChatMessage::User { content, .. } = message {
-                if let UserMessageContent::Parts(parts) = content {
-                    for part in parts {
-                        if let ContentPart::ImageUrl { image_url } = part {
-                            image_urls.push(image_url.url.clone());
-                        }
-                    }
-                }
-            }
-        }
-
-        if image_urls.is_empty() && video_urls.is_empty() && audio_urls.is_empty() {
-            Ok(None)
-        } else {
-            Ok(Some(proto::MultimodalInputs {
-                image_urls,
-                video_urls,
-                audio_urls,
-                ..Default::default()
-            }))
-        }
-    }
-
-    /// Build gRPC SamplingParams from OpenAI request
-    fn build_grpc_sampling_params(
-        &self,
-        request: &ChatCompletionRequest,
-        structural_tag: Option<String>,
-    ) -> proto::SamplingParams {
-        // Extract stop sequences
-        let stop_sequences = self.extract_stop_strings(request);
-
-        proto::SamplingParams {
-            temperature: request.temperature.unwrap_or(1.0),
-            top_p: request.top_p.unwrap_or(1.0),
-            top_k: request.top_k.unwrap_or(-1) as i32,
-            min_p: request.min_p.unwrap_or(0.0),
-            frequency_penalty: request.frequency_penalty.unwrap_or(0.0),
-            presence_penalty: request.presence_penalty.unwrap_or(0.0),
-            repetition_penalty: request.repetition_penalty.unwrap_or(1.0),
-            max_new_tokens: request.max_tokens.unwrap_or(128) as i32,
-            stop: stop_sequences,
-            stop_token_ids: request.stop_token_ids.clone().unwrap_or_default(),
-            skip_special_tokens: request.skip_special_tokens,
-            n: request.n.unwrap_or(1) as i32,
-            structural_tag: structural_tag.unwrap_or_default(),
-            // Handle constraints
-            constraint: self.build_constraint(request),
-            ..Default::default()
-        }
-    }
-
-    /// Extract stop strings from request
-    fn extract_stop_strings(&self, request: &ChatCompletionRequest) -> Vec<String> {
-        match &request.stop {
-            Some(StringOrArray::String(s)) => vec![s.clone()],
-            Some(StringOrArray::Array(arr)) => arr.clone(),
-            None => vec![],
-        }
-    }
-
-    /// Build constraint for structured generation
-    fn build_constraint(&self, request: &ChatCompletionRequest) -> Option<proto::sampling_params::Constraint> {
-        // Check for response_format
-        if let Some(format) = &request.response_format {
-            if let ResponseFormat::JsonSchema { json_schema } = format {
-                return Some(proto::sampling_params::Constraint::JsonSchema(
-                    serde_json::to_string(&json_schema.schema).unwrap_or_default()
-                ));
-            }
-        }
-
-        // Check for ebnf grammar
-        if let Some(ebnf) = &request.ebnf {
-            return Some(proto::sampling_params::Constraint::EbnfGrammar(ebnf.clone()));
-        }
-
-        // Check for regex
-        if let Some(regex) = &request.regex {
-            return Some(proto::sampling_params::Constraint::Regex(regex.clone()));
-        }
-
-        None
-    }
-
-    /// Generate tool constraints for structured generation
-    fn generate_tool_constraints(
-        &self,
-        _tools: &[crate::protocols::spec::Tool],
-        _tool_choice: &Option<crate::protocols::spec::ToolChoice>,
-        model: &str,
-    ) -> Option<String> {
-        // Get the appropriate tool parser for the model
-        let _parser = self.tool_parser_registry.get_parser(model)?;
-
-        // Generate structural tags for constrained generation
-        // This would call the parser's method to generate constraints
-        // For now, returning None as the actual implementation depends on the parser
-        None
-    }
-
-    /// Select a worker for the request
-    fn select_worker_for_request(
-        &self,
-        model_id: Option<&str>,
-        text: &str,
-    ) -> Option<Arc<dyn Worker>> {
-        let workers = self.workers.read().unwrap();
-
-        // Filter workers by model if specified
-        let available: Vec<_> = workers
-            .iter()
-            .filter(|w| {
-                // Check if worker is healthy and circuit allows execution
-                if !w.is_healthy() || !w.circuit_breaker().can_execute() {
-                    return false;
-                }
-
-                // Check model match if specified
-                if let Some(_model) = model_id {
-                    // Worker should match the model
-                    // This is simplified - actual implementation would check worker's model
-                    true
-                } else {
-                    true
-                }
-            })
-            .cloned()
-            .collect();
-
-        if available.is_empty() {
-            return None;
-        }
-
-        // Use load balancing policy to select worker
-        if let Some(idx) = self.policy.select_worker(&available, Some(text)) {
-            Some(available[idx].clone())
-        } else {
-            warn!("Policy failed to select worker");
-            // Fallback to first available worker
-            available.first().cloned()
-        }
-    }
-
-    /// Get or create a gRPC client for the worker
-    async fn get_or_create_grpc_client(
-        &self,
-        worker_url: &str,
-    ) -> Result<SglangSchedulerClient, String> {
-        // For now, create a new client each time since SglangSchedulerClient doesn't implement Clone
-        // In production, we'd want to implement a connection pool
-        debug!("Creating new gRPC client for worker: {}", worker_url);
-        SglangSchedulerClient::connect(worker_url)
-            .await
-            .map_err(|e| format!("Failed to connect to gRPC server: {}", e))
-    }
-
-    /// Placeholder for streaming handler (to be implemented in Phase 2)
-    async fn handle_streaming_chat(
-        &self,
-        _client: SglangSchedulerClient,
-        _request: proto::GenerateRequest,
-        _original_request: &ChatCompletionRequest,
-    ) -> Response {
-        // This will be implemented in Phase 2
-        (StatusCode::NOT_IMPLEMENTED, "Streaming not yet implemented").into_response()
-    }
-
-    /// Placeholder for non-streaming handler (to be implemented in Phase 3)
-    async fn handle_non_streaming_chat(
-        &self,
-        _client: SglangSchedulerClient,
-        _request: proto::GenerateRequest,
-        _original_request: &ChatCompletionRequest,
-    ) -> Response {
-        // This will be implemented in Phase 3
-        (StatusCode::NOT_IMPLEMENTED, "Non-streaming not yet implemented").into_response()
-    }
-}
