--- sgl-router/src/routers/grpc/handlers/mod.rs
@@ -1,3 +1,0 @@
-// gRPC route handlers module
-
-pub mod chat;
