+++ sgl-router/src/routers/grpc/mod.rs
@@ -1,5 +1,4 @@
 //! gRPC router implementations
 
-pub mod handlers;
 pub mod pd_router;
 pub mod router;
